* Fixing #171 Be able to select encoder before selecting video
