+++ fastflix/conversion_worker.py
@@ -246,6 +246,7 @@ def queue_worker(gui_proc, worker_queue, status_queue, log_queue, queue_list, qu
                 currently_encoding = False
                 allow_sleep_mode()
                 status_queue.put(("cancelled", video.uuid if video else ""))
+                log_queue.put("STOP_TIMER")
 
             if request[0] == "pause queue":
                 logger.debug(t("Command worker received request to pause encoding after the current item completes"))
